style: Organize & rename components
