+++ src/lib/components/Icon.svelte
