+++ src/lib/components/InfoIcon.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import type { SvelteHTMLElements } from "svelte/elements";
-    import Icon from "$lib/components/IconLoader.svelte";
+    import Icon from "$lib/components/Icon.svelte";
 
     let { children }: SvelteHTMLElements["span"] = $props();
 </script>
