+++ src/lib/components/bento/BadgeMarquee.svelte
